+++ React Project/src/component/SuggestionPage.js
@@ -1,0 +1,9 @@
+function SuggestionPage() {
+  return (
+    <div>
+      <h1>HI</h1>
+    </div>
+  );
+}
+
+export default SuggestionPage;
